feat: renamed rtk query file for boards and updated imports
